requirements.txt
discord.py
wavelink>=1.4.0
python-dotenv